refactor: fix double exist error and extract color option
