+++ lib/src/editor/toolbar/util/color_option.dart
@@ -1,0 +1,10 @@
+class ColorOption {
+  const ColorOption({
+    required this.colorHex,
+    required this.name,
+  });
+
+  // 0xFF000000
+  final String colorHex;
+  final String name;
+}
