+++ lib/src/editor/toolbar/util/util.dart
@@ -1,1 +1,5 @@
 export 'color_generators.dart';
+export 'color_option.dart';
+// formatHighlightColor & formatFontColor
+// also exist in package:appflowy_editor/src/service/default_text_operations/format_rich_text_style.dart
+// export 'format_color.dart';
